requirements.txt
asgiref==3.3.1
Django==3.1.7
django-autoslug==1.9.8
django-ckeditor==6.1.0
django-js-asset==1.2.2
django-taggit==1.5.1
Pillow==8.2.0
python-dotenv==0.19.0
pytz==2021.1
sqlparse==0.4.1
